@@ -1,13 +1,9 @@
 FactoryGirl.define do
   factory :user do
     email
-
-    full_name 'John Smith'
-    password '123456'
-    password_confirmation '123456'
-  end
-
-  trait :confirmed do
+    full_name { Faker::Name.name }
+    password "123456"
+    password_confirmation { password }
     confirmed_at 1.hour.ago
   end
 
@@ -15,7 +11,7 @@ FactoryGirl.define do
     confirmed_at nil
 
     after(:create) do |user|
-      user.update_attributes(confirmation_sent_at: 3.days.ago)
+      user.update(confirmation_sent_at: 3.days.ago)
     end
   end
 end
