@@ -1,15 +1,11 @@
 class ApplicationController < ActionController::Base
-  protect_from_forgery
+  include Authentication
+  include Authorization
 
-  protected
+  protect_from_forgery with: :exception
 
-  def devise_parameter_sanitizer
-    if resource_class == User
-      User::ParameterSanitizer.new(User, :user, params)
-    else
-      super
-    end
-  end
+  responders :flash
+  respond_to :html
 
   decent_configuration do
     strategy DecentExposure::StrongParametersStrategy
