@@ -30,9 +30,9 @@ Rails.application.configure do
   # number of complex assets.
   config.assets.debug = true
 
-  # Application specific options
-  #
-  config.slim_options =  { pretty: true }
+  # Asset digests allow you to set far-future HTTP expiration dates on all assets,
+  # yet still be able to expire them through the digest params.
+  config.assets.digest = true
 
   # Adds additional error checking when serving assets at runtime.
   # Checks for improperly declared sprockets dependencies.
