@@ -1,1 +1,1 @@
-Slim::Engine.set_default_options(app_config.slim_options)
+Slim::Engine.set_options(pretty: ENV.fetch("SLIM_PRETTY", false))
