@@ -1,4 +1,4 @@
-require 'rails_helper'
+require "rails_helper"
 
 describe User do
   it { is_expected.to validate_presence_of :full_name }
