@@ -1,25 +1,13 @@
-ENV['RAILS_ENV'] ||= 'test'
-require 'spec_helper'
-require File.expand_path('../../config/environment', __FILE__)
-require 'rspec/rails'
+ENV["RAILS_ENV"] ||= "test"
 
-# Run codeclimate-test-reporter only in CI
-if ENV['CI']
-  require 'codeclimate-test-reporter'
-  CodeClimate::TestReporter.start
-end
+require "spec_helper"
+require File.expand_path("../../config/environment", __FILE__)
+require "rspec/rails"
+require "shoulda/matchers"
 
-Dir[Rails.root.join('spec/support/**/*.rb')].each { |f| require f }
+Dir[Rails.root.join("spec/support/**/*.rb")].each { |f| require f }
 
 RSpec.configure do |config|
-  config.use_transactional_fixtures = true
+  config.use_transactional_fixtures = false
   config.infer_spec_type_from_file_location!
-
-  config.include EmailSpec::Helpers
-  config.include EmailSpec::Matchers
-  config.include FactoryGirl::Syntax::Methods
-
-  config.before do
-    ActionMailer::Base.deliveries.clear
-  end
 end
